requirements.txt
PyYAML==6.0.1
typer==0.9.0
requests==2.31.0
beautifulsoup4==4.12.2
pytest==7.4.3
python-dotenv==1.0.0 